Implementing data store (base, fs)
