+++ backend/api/user/router.py
@@ -12,7 +12,7 @@ import models.api as api
 import models.core as core
 from models.enums import ActionType, RequestType, ResourceType, ContentType
 from utils.custom_validations import validate_uniqueness
-import utils.db as db
+from utils.data_database import data_adapter as db
 from utils.access_control import access_control
 from utils.helpers import flatten_dict
 from utils.custom_validations import validate_payload_with_schema
