+++ backend/plugins/redis_db_update/plugin.py
@@ -3,7 +3,7 @@ from models.core import ActionType, Attachment, PluginBase, Event, Space
 from utils.helpers import camel_case
 from utils.repository import generate_payload_string
 from utils.spaces import get_spaces
-import utils.db as db
+from utils.data_database import data_adapter as db
 from models import core
 from models.enums import ContentType, ResourceType
 from utils.redis_services import RedisServices
